Refactor sanitized_integer method
All the logic to sanitize either an absolute day reference or an apocalypse
reckoning reference, and raise an appropriate error where necessary, has been
placed into the sanitized_integer method so that it can be used whenever and
wherever a CalmarendianDate object is constructed or updated with a day
reference. Also a (rather serious) error, whereby the date elements were not
updated when adr (absolute day reference) was changed has been fixed.

This rather means that both v0.1.1 and v0.2.0 are broken.