Implement the today() method
There is no real meaning to the term 'today' as a reference to a Calmarendian
date from the perspective of an observer on Earth. The method name has been
chosen (a) to mirror the method of the same name in Python's Gregorian date
class and (b) because, from our world-building perspective, everything that
happened before Jennifer's arrival is in the past, everything that happened
afterwards is deemed the future implying that 777-7-03-1 must be 'today'.
However, unlike the standard Python date.today() method, this one will always
return the same value whenever it is invoked.