Add getters and setters to CalmarendianTimeDelta
Boilerplate for getters and setters added for days, seconds and microseconds;
full type and error checking are still to be added on the setters. It has been
decided that, by design, CalmarendianTimeDelta objects should not be entirely
immutable.

A __str__ method added whose functionality closely replicates that of the
Gregorian timedelta class.

Expansion of the unit tests indicates that the constructor works correctly with
no arguments and with a wider range of integer and float day values other than
just 1.